refactor: split session-thread into single-component files
react-doctor flagged six no-multi-comp warnings in session-thread.tsx (seven
components in one file). Extract AssistantMessage (+ its private error
marker), ScrollToBottomPill, ThreadViewport, and ThreadMessages (+ its row
list) into their own files, leaving SessionThread as the composition root.
Changed-scope score goes 98 -> 100 with no behavior change.
