+++ web/src/components/assistant-ui/session-thread.tsx
@@ -1,42 +1,17 @@
-import {
-  MessagePrimitive,
-  ThreadPrimitive,
-  ReadonlyThreadProvider,
-  type ThreadMessage,
-  useAui,
-  useAuiState,
-} from "@assistant-ui/react";
-import { AlertCircle, ArrowDown } from "lucide-react";
-import { type ComponentPropsWithoutRef, useEffect, useLayoutEffect, useRef } from "react";
+import { ThreadPrimitive, useAui } from "@assistant-ui/react";
 
-import { cn } from "@/lib/utils";
-import { Marker } from "@compozy/ui";
 import { SessionGoalHeaderContainer } from "@/systems/session/components/goal/session-goal-header-container";
 import { SessionGoalCommandErrorNotice } from "@/systems/session/components/goal/session-goal-command-error-notice";
-import { SessionLoadOlderButton } from "@/systems/session/components/session-load-older-button";
-import { SessionDecisionDock, useSessionTranscriptThreadState } from "@/systems/session";
+import { SessionDecisionDock } from "@/systems/session";
 import type { SessionFailurePayload, SessionState } from "@/systems/session";
-import { isGoalCommandFailureGuidance } from "@/systems/session/lib/session-goal-chat-transport";
-import {
-  recordSessionDebugEvent,
-  SESSION_DEBUG_EVENTS,
-} from "@/systems/session/lib/session-observability";
+import { useSessionComposerState } from "./hooks/use-session-composer-state";
 import { SessionComposer, type SessionComposerProps } from "./session-composer";
+import { SessionComposerPrefillProvider } from "./session-composer-prefill-context";
 import {
   SESSION_THREAD_CONTENT_INSET_DEFAULT,
   ThreadContentRail,
-  type SessionThreadContentInset,
 } from "./session-thread-content-rail";
-import { threadProviderIdentity } from "./hooks/use-thread-provider-identity";
-import { useThreadScrollController } from "./hooks/use-thread-scroll-controller";
-import { useSessionComposerState } from "./hooks/use-session-composer-state";
-import { MessageActions } from "./message-actions";
-import { SessionComposerPrefillProvider } from "./session-composer-prefill-context";
-import { formatMessageError } from "./session-thread-error";
-import { UserMessage } from "./session-user-message";
-import { SessionThreadErrorBoundary } from "./session-thread-error-boundary";
-import { AssistantMessageTimeline } from "./session-timeline-render";
-import { ThreadStatePane } from "./session-thread-states";
+import { ThreadViewport } from "./session-thread-viewport";
 
 const EMPTY_QUEUED_PROMPTS: NonNullable<SessionComposerProps["queuedPrompts"]> = [];
 
@@ -49,296 +24,11 @@ interface SessionThreadProps extends SessionComposerProps {
   failure?: SessionFailurePayload | null;
 }
 
-function SessionMessageErrorNotice() {
-  const error = useAuiState(state => {
-    const status = state.message.status;
-    if (status?.type !== "incomplete" || status.reason !== "error") {
-      return null;
-    }
-    return formatMessageError(status.error);
-  });
-
-  if (error === null || isGoalCommandFailureGuidance(error)) {
-    return null;
-  }
-
-  return (
-    <Marker
-      role="alert"
-      data-testid="session-message-error"
-      tone="danger"
-      icon={<AlertCircle strokeWidth={1.8} />}
-    >
-      {error}
-    </Marker>
-  );
-}
-
-function AssistantMessage() {
-  const isRunning = useAuiState(state => {
-    const status = state.message.status;
-    return status?.type === "running" || status?.type === "incomplete";
-  });
-
-  return (
-    <MessagePrimitive.Root
-      className={cn("group/message flex w-full min-w-0 pt-1", isRunning ? "pb-[7px]" : "pb-[18px]")}
-    >
-      <div className="flex min-w-0 flex-1 flex-col gap-[7px]">
-        <SessionThreadErrorBoundary>
-          <AssistantMessageTimeline />
-        </SessionThreadErrorBoundary>
-        <SessionMessageErrorNotice />
-        <MessageActions
-          align="start"
-          copyLabel="Copy message"
-          goalPrefill
-          testId="assistant-message-actions"
-        />
-      </div>
-    </MessagePrimitive.Root>
-  );
-}
-
-type ThreadViewportProps = ComponentPropsWithoutRef<typeof ThreadPrimitive.Viewport>;
-
-export function ScrollToBottomPill({
-  visible,
-  onClick,
-}: {
-  visible: boolean;
-  onClick: () => void;
-}) {
-  // Floating scroll-to-bottom affordance: a neutral `size-8` disc (no
-  // glass/backdrop-blur) that fades + drifts in with the shared disclosure
-  // motion and stays mounted so its exit animates too.
-  return (
-    <div
-      className={cn(
-        "pointer-events-none absolute inset-x-0 bottom-3 z-20 flex justify-center",
-        "transition-[transform,opacity] duration-base ease-out motion-reduce:transition-none",
-        visible ? "translate-y-0 opacity-100" : "translate-y-1 opacity-0"
-      )}
-      aria-hidden={!visible}
-    >
-      <button
-        type="button"
-        data-testid="scroll-to-bottom-pill"
-        data-visible={visible}
-        aria-label="Scroll to latest"
-        onClick={onClick}
-        tabIndex={visible ? 0 : -1}
-        className={cn(
-          "flex size-8 items-center justify-center rounded-full",
-          "border border-line bg-canvas-soft text-muted shadow-[var(--shadow-overlay)]",
-          "transition-colors hover:bg-hover hover:text-fg",
-          "focus-visible:outline-none focus-visible:shadow-focus-ring",
-          visible ? "pointer-events-auto" : "pointer-events-none"
-        )}
-      >
-        <ArrowDown className="size-3.5" aria-hidden="true" />
-      </button>
-    </div>
-  );
-}
-
-function ThreadViewport({
-  agentName,
-  sessionId,
-  isSessionRunning,
-  contentInset,
-  sessionState,
-  failure,
-  startupFailed,
-  className,
-  ...props
-}: ThreadViewportProps & {
-  agentName: string;
-  sessionId: string;
-  isSessionRunning: boolean;
-  contentInset: SessionThreadContentInset;
-  sessionState?: SessionState;
-  failure?: SessionFailurePayload | null;
-  startupFailed: boolean;
-}) {
-  const viewportRef = useRef<HTMLDivElement | null>(null);
-  const {
-    messages: transcriptMessages,
-    status: transcriptStatus,
-    error: transcriptError,
-    hasOlder,
-    isFetchingOlder,
-    loadOlder,
-    retry: retryTranscript,
-  } = useSessionTranscriptThreadState();
-  const olderAnchorRef = useRef<{ messageId: string; offsetTop: number } | null>(null);
-  const messageCount = transcriptMessages.length;
-  const { showScrollToBottom, scrollToEnd } = useThreadScrollController(
-    viewportRef,
-    transcriptMessages
-  );
-  const loadOlderWithAnchor = () => {
-    const viewport = viewportRef.current;
-    if (viewport) {
-      const viewportTop = viewport.getBoundingClientRect().top;
-      const messageRows = [...viewport.querySelectorAll<HTMLElement>("[data-message-id]")];
-      const anchorRow =
-        messageRows.find(row => row.getBoundingClientRect().bottom >= viewportTop) ??
-        messageRows[0];
-      const messageId = anchorRow?.dataset.messageId;
-      olderAnchorRef.current =
-        anchorRow && messageId
-          ? { messageId, offsetTop: anchorRow.getBoundingClientRect().top - viewportTop }
-          : null;
-    }
-    loadOlder();
-  };
-
-  useLayoutEffect(() => {
-    const viewport = viewportRef.current;
-    const anchor = olderAnchorRef.current;
-    if (!viewport || !anchor || isFetchingOlder) return;
-    const anchorRow = [...viewport.querySelectorAll<HTMLElement>("[data-message-id]")].find(
-      row => row.dataset.messageId === anchor.messageId
-    );
-    if (anchorRow) {
-      const nextOffset =
-        anchorRow.getBoundingClientRect().top - viewport.getBoundingClientRect().top;
-      viewport.scrollTop += nextOffset - anchor.offsetTop;
-    }
-    olderAnchorRef.current = null;
-  }, [isFetchingOlder, messageCount]);
-
-  return (
-    <div className="relative flex min-h-0 min-w-0 flex-1 flex-col">
-      <ThreadPrimitive.Viewport
-        {...props}
-        ref={viewportRef}
-        className={cn("min-h-0 flex-1 overflow-y-auto", className)}
-        data-testid="chat-view"
-      >
-        <ThreadContentRail inset={contentInset} className="min-h-full">
-          {hasOlder || isFetchingOlder ? (
-            <SessionLoadOlderButton isLoading={isFetchingOlder} onClick={loadOlderWithAnchor} />
-          ) : null}
-          <ThreadMessages
-            agentName={agentName}
-            sessionId={sessionId}
-            isSessionRunning={isSessionRunning}
-            messageCount={messageCount}
-            transcriptStatus={transcriptStatus}
-            transcriptError={transcriptError}
-            retryTranscript={retryTranscript}
-            transcriptMessages={transcriptMessages}
-            sessionState={sessionState}
-            failure={failure}
-            startupFailed={startupFailed}
-          />
-        </ThreadContentRail>
-      </ThreadPrimitive.Viewport>
-      <ScrollToBottomPill visible={showScrollToBottom} onClick={scrollToEnd} />
-    </div>
-  );
-}
-
-const SESSION_MESSAGE_COMPONENTS = {
-  UserMessage,
-  AssistantMessage,
-};
-
-function ThreadMessageRows({
-  messageCount,
-  transcriptMessages,
-}: {
-  messageCount: number;
-  transcriptMessages: readonly ThreadMessage[];
-}) {
-  const committedMessageCount = useAuiState(state => state.thread.messages.length);
-  const renderableCount = Math.min(messageCount, committedMessageCount);
-
-  return (
-    <div className="w-full" data-testid="thread-messages">
-      {Array.from({ length: renderableCount }, (_, index) => (
-        <div
-          key={transcriptMessages[index]?.id ?? index}
-          data-message-id={transcriptMessages[index]?.id}
-          data-message-index={index}
-          data-testid="thread-message-row"
-          className="w-full [content-visibility:auto] [contain-intrinsic-size:auto_120px]"
-        >
-          <ThreadPrimitive.MessageByIndex index={index} components={SESSION_MESSAGE_COMPONENTS} />
-        </div>
-      ))}
-    </div>
-  );
-}
-
-export function ThreadMessages({
-  agentName,
-  sessionId,
-  isSessionRunning,
-  messageCount,
-  transcriptStatus,
-  transcriptError,
-  retryTranscript,
-  transcriptMessages,
-  sessionState,
-  failure,
-  startupFailed,
-}: {
-  agentName: string;
-  sessionId: string;
-  isSessionRunning: boolean;
-  messageCount: number;
-  transcriptStatus: ReturnType<typeof useSessionTranscriptThreadState>["status"];
-  transcriptError: ReturnType<typeof useSessionTranscriptThreadState>["error"];
-  retryTranscript: () => void;
-  transcriptMessages: ReturnType<typeof useSessionTranscriptThreadState>["messages"];
-  sessionState?: SessionState;
-  failure?: SessionFailurePayload | null;
-  startupFailed: boolean;
-}) {
-  const emptyWhileActive = messageCount === 0 && transcriptStatus === "success" && isSessionRunning;
-  useEffect(() => {
-    if (!emptyWhileActive) {
-      return;
-    }
-    recordSessionDebugEvent(SESSION_DEBUG_EVENTS.threadEmptyWhileActive, {
-      agent_name: agentName,
-      message_count: messageCount,
-      session_id: sessionId,
-      transcript_status: transcriptStatus,
-    });
-  }, [agentName, emptyWhileActive, messageCount, sessionId, transcriptStatus]);
-
-  const transcriptIdentity = threadProviderIdentity(transcriptMessages);
-
-  if (messageCount === 0) {
-    return (
-      <ThreadStatePane
-        status={transcriptStatus}
-        agentName={agentName}
-        error={transcriptError}
-        onRetry={retryTranscript}
-        sessionState={sessionState}
-        failure={failure}
-        startupFailed={startupFailed}
-      />
-    );
-  }
-
-  // `ReadonlyThreadProvider` applies changed `messages` in a passive effect, but
-  // appended/removed message ids change which indexes are valid immediately.
-  // Keying only on the id sequence rebuilds that provider with the correct core
-  // size synchronously while leaving the outer virtualizer and its scroll machine
-  // alive across the remount.
-  return (
-    <ReadonlyThreadProvider key={transcriptIdentity} messages={transcriptMessages}>
-      <ThreadMessageRows messageCount={messageCount} transcriptMessages={transcriptMessages} />
-    </ReadonlyThreadProvider>
-  );
-}
-
+/**
+ * The session surface composition root: pinned goal zone above the transcript
+ * scroller, the viewport itself, and the composer zone (goal-command notice,
+ * decision dock, composer) below it.
+ */
 export function SessionThread({
   sessionId,
   agentName,
