+++ web/src/systems/session/components/assistant-ui/stories/session-thread.stories.tsx
@@ -2,7 +2,8 @@ import { primarySessionFixture } from "@/systems/session/mocks";
 import { type ComponentProps, type ReactNode, useEffect } from "react";
 import { sessionStore } from "@/systems/session/stores/session-store";
 import type { Meta, StoryObj } from "@storybook/react-vite";
-import { ScrollToBottomPill, SessionThread } from "@/components/assistant-ui/session-thread";
+import { ScrollToBottomPill } from "@/components/assistant-ui/scroll-to-bottom-pill";
+import { SessionThread } from "@/components/assistant-ui/session-thread";
 import { storybookMswParameters } from "@/storybook/msw";
 import { compozyApiMock } from "@/storybook/openapi-msw";
 import { HttpResponse } from "msw";
